test: split `test_converters.py` into `tests/converters/` and `tests/applicators/` mirroring source
