+++ tests/applicators/__init__.py
@@ -1,0 +1,3 @@
+"""Tests for the applicators package."""
+
+__all__: list[str] = []
