+++ tests/applicators/test_one_of.py
