+++ tests/converters/__init__.py
@@ -1,0 +1,3 @@
+"""Tests for the converters package."""
+
+__all__: list[str] = []
